add telegram link and optimize components
