+++ src/components/SideNav/index.js
@@ -5,13 +5,12 @@ import Title from '../Title'
 import { BasicLink } from '../Link'
 import { useMedia } from 'react-use'
 import { transparentize } from 'polished'
-import { TYPE } from '../../Theme'
 import { withRouter } from 'react-router-dom'
 import { TrendingUp, List, PieChart, Disc } from 'react-feather'
 import Link from '../Link'
-import { useSessionStart } from '../../contexts/Application'
 import { useDarkModeManager } from '../../contexts/LocalStorage'
 import Toggle from '../Toggle'
+import SideNavTimer from '../SideNavTimer'
 
 const Wrapper = styled.div`
   height: ${({ isMobile }) => (isMobile ? 'initial' : '100vh')};
@@ -75,36 +74,8 @@ const HeaderText = styled.div`
   }
 `
 
-const Polling = styled.div`
-  position: fixed;
-  display: flex;
-  left: 0;
-  bottom: 0;
-  padding: 1rem;
-  color: white;
-  opacity: 0.4;
-  transition: opacity 0.25s ease;
-  :hover {
-    opacity: 1;
-  }
-`
-const PollingDot = styled.div`
-  width: 8px;
-  height: 8px;
-  min-height: 8px;
-  min-width: 8px;
-  margin-right: 0.5rem;
-  margin-top: 3px;
-  border-radius: 50%;
-  background-color: ${({ theme }) => theme.green1};
-`
-
 function SideNav({ history }) {
   const below1080 = useMedia('(max-width: 1080px)')
-
-  const below1180 = useMedia('(max-width: 1180px)')
-
-  const seconds = useSessionStart()
 
   const [isDark, toggleDarkMode] = useDarkModeManager()
 
@@ -183,18 +154,14 @@ function SideNav({ history }) {
                 Twitter
               </Link>
             </HeaderText>
+            <HeaderText>
+              <Link href="https://t.me/fuseswap" target="_blank">
+                Telegram
+              </Link>
+            </HeaderText>
             <Toggle isActive={isDark} toggle={toggleDarkMode} />
           </AutoColumn>
-          {!below1180 && (
-            <Polling style={{ marginLeft: '.5rem' }}>
-              <PollingDot />
-              <a href="/" style={{ color: 'white' }}>
-                <TYPE.small color={'white'}>
-                  Updated {!!seconds ? seconds + 's' : '-'} ago <br />
-                </TYPE.small>
-              </a>
-            </Polling>
-          )}
+          <SideNavTimer />
         </DesktopWrapper>
       ) : (
         <MobileWrapper>
